# fix(path): fix common() results that depend on the order and/or number of input paths